requirements.txt
anthropic>=0.34.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langgraph>=0.2.0
langsmith>=0.1.0
langgraph-cli[inmem]>=0.2.0

